fixed the client/server component stuff
